feat: Update Discord bot to use interactionCreate event for grammar fixing command
This commit modifies the Discord bot script to use the interactionCreate event instead of the messageCreate event for the grammar fixing command. It also updates the command to accept options and retrieves the user's message from the options. This change improves the functionality and user experience of the Discord bot by utilizing the latest Discord API features.